+++ Kokoro/ModEntry.cs
@@ -3,24 +3,22 @@ using CobaltCoreModding.Definitions.ModContactPoints;
 using CobaltCoreModding.Definitions.ModManifests;
 using HarmonyLib;
 using Microsoft.Extensions.Logging;
-using Nanoray.Pintail;
+using Nanoray.PluginManager;
 using Newtonsoft.Json.Serialization;
-using Shockah.Shared;
+using Nickel.Legacy;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Reflection;
-using System.Reflection.Emit;
 
 namespace Shockah.Kokoro;
 
-public sealed class ModEntry : IModManifest, IPrelaunchManifest, IApiProviderManifest, ISpriteManifest, IStatusManifest
+public sealed class ModEntry : IModManifest, IPrelaunchManifest, IApiProviderManifest, ISpriteManifest, IStatusManifest, INickelManifest
 {
 	internal const string ExtensionDataJsonKey = "KokoroModData";
 	internal const string ScorchingTag = "Scorching";
 
 	public static ModEntry Instance { get; private set; } = null!;
-	internal IProxyManager<string> ProxyManager { get; private set; } = null!;
+	internal Nickel.IModHelper Helper { get; private set; } = null!;
 	internal ApiImplementation Api { get; private set; } = null!;
 	private Harmony Harmony = null!;
 
@@ -31,35 +29,26 @@ public sealed class ModEntry : IModManifest, IPrelaunchManifest, IApiProviderMan
 	public DirectoryInfo? ModRootFolder { get; set; }
 	public ILogger? Logger { get; set; }
 
-	internal Content Content = new();
-	internal ExtensionDataManager ExtensionDataManager { get; private init; } = new();
-	public EvadeManager EvadeManager { get; private init; } = new();
-	public DroneShiftManager DroneShiftManager { get; private init; } = new();
-	public ArtifactIconManager ArtifactIconManager { get; private init; } = new();
-	public StatusRenderManager StatusRenderManager { get; private init; } = new();
-	public StatusLogicManager StatusLogicManager { get; private init; } = new();
-	public CardRenderManager CardRenderManager { get; private init; } = new();
-	public WrappedActionManager WrappedActionManager { get; private init; } = new();
-	public MidrowScorchingManager MidrowScorchingManager { get; private init; } = new();
-	public WormStatusManager WormStatusManager { get; private init; } = new();
-	public OxidationStatusManager OxidationStatusManager { get; private init; } = new();
-	public RedrawStatusManager RedrawStatusManager { get; private init; } = new();
-	public StatusNextTurnManager StatusNextTurnManager { get; private init; } = new();
+	internal readonly Content Content = new();
+	internal readonly ExtensionDataManager ExtensionDataManager = new();
+	public readonly EvadeManager EvadeManager = new();
+	public readonly DroneShiftManager DroneShiftManager = new();
+	public readonly ArtifactIconManager ArtifactIconManager = new();
+	public readonly StatusRenderManager StatusRenderManager = new();
+	public readonly StatusLogicManager StatusLogicManager = new();
+	public readonly CardRenderManager CardRenderManager = new();
+	public readonly WrappedActionManager WrappedActionManager = new();
+	public readonly MidrowScorchingManager MidrowScorchingManager = new();
+	public readonly WormStatusManager WormStatusManager = new();
+	public readonly OxidationStatusManager OxidationStatusManager = new();
+	public readonly RedrawStatusManager RedrawStatusManager = new();
+	public readonly StatusNextTurnManager StatusNextTurnManager = new();
 
 	internal TimeSpan TotalGameTime;
 
 	public void BootMod(IModLoaderContact contact)
 	{
 		Instance = this;
-		ReflectionExt.CurrentAssemblyLoadContext.LoadFromAssemblyPath(Path.Combine(ModRootFolder!.FullName, "Pintail.dll"));
-
-		AssemblyBuilder assemblyBuilder = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName($"CobaltCoreModding.Proxies, Version={this.GetType().Assembly.GetName().Version}, Culture=neutral"), AssemblyBuilderAccess.Run);
-		ModuleBuilder moduleBuilder = assemblyBuilder.DefineDynamicModule("CobaltCoreModding.Proxies");
-		ProxyManager = new ProxyManager<string>(moduleBuilder, new ProxyManagerConfiguration<string>(
-			proxyPrepareBehavior: ProxyManagerProxyPrepareBehavior.Eager,
-			proxyObjectInterfaceMarking: ProxyObjectInterfaceMarking.MarkerWithProperty,
-			accessLevelChecking: AccessLevelChecking.DisabledButOnlyAllowPublicMembers
-		));
 		Api = new(this);
 		Api.RegisterTypeForExtensionData(typeof(ACardOffering));
 		Api.RegisterTypeForExtensionData(typeof(AStatus));
@@ -100,10 +89,11 @@ public sealed class ModEntry : IModManifest, IPrelaunchManifest, IApiProviderMan
 		SetupSerializationChanges();
 	}
 
+	public void OnNickelLoad(IPluginPackage<Nickel.IModManifest> package, Nickel.IModHelper helper)
+		=> Helper = helper;
+
 	public void FinalizePreperations(IPrelaunchContactPoint prelaunchManifest)
-	{
-		StuffBasePatches.ApplyLate(Harmony);
-	}
+		=> StuffBasePatches.ApplyLate(Harmony);
 
 	public object? GetApi(IManifest requestingMod)
 		=> new ApiImplementation(requestingMod);