+++ Kokoro/docs/release-notes.md
@@ -2,7 +2,8 @@
 
 # Release notes
 
-## Upcoming release
+## 1.11.2
+Released 14 July 2024.
 
 * The Oxidation status now flashes when about to turn into Corrode.
 * Switched to using Nickel's newly exposed proxy manager, potentially improving performance and compatibility.